chore(autoforager): Project restructure
