+++ src/AutoForager/Extensions/TreeExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using StardewValley.TerrainFeatures;
 using System.Collections.Generic;
 
-namespace AutoForager.Helpers
+namespace AutoForager.Extensions
 {
     public static class TreeExtensions
     {
